Fix: 레코드 검색 타입 변경
